code: restore IosDriver per request
